+++ gdsfactory/types.py
@@ -90,6 +90,8 @@ CrossSectionOrFactory = Union[CrossSection, Callable[..., CrossSection]]
 PortSymmetries = Dict[str, Dict[str, List[str]]]
 
 ComponentSpec = Union[str, ComponentFactory, Component, Dict[str, Any]]
+CellSpec = Union[str, ComponentFactory, Dict[str, Any]]
+ComponentSpecOrList = Union[ComponentSpec, List[ComponentSpec]]
 CrossSectionSpec = Union[str, CrossSectionFactory, CrossSection, Dict[str, Any]]