+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # [CHANGELOG](https://keepachangelog.com/en/1.0.0/)
 
+## [5.0.2]
+
+- get_cells and get_component_factories work with module and list of modules
+- add `gf.get_cells` and `gf.get_cross_section_factories`
+- get_component and get_cross_section accepts also omegaconf.DictConfig
+- add pack_doe and pack_doe_grid to containers
+- add gf.get_cell, and enable partials
+
 ## [5.0.1](https://github.com/gdsfactory/gdsfactory/pull/274)
 
 - fix bends bbox
-
 
 ## [5.0.0](https://github.com/gdsfactory/gdsfactory/pull/273)